Lift tapAtImpl into InteractiveBase so every renderer shares one copy
